+++ lib/schema.json
@@ -29,7 +29,6 @@
 
         "entity": {
             "type": "object",
-            "minProperties": 1,
             "additionalProperties": false,
             "properties": {
                 "block": { "type": "string" },
